Add debug options to doc-builder

+++ make_docs_without_docker.sh
@@ -1,6 +1,48 @@
 #!/bin/bash
 # file to setup documentation building pipeline without docker
 # $1 : Path to the library being documented
+# --no-cleanup disable the backup/cleanup procedure
+# --git-add stage changed files before generating the docs
+
+PROGNAME=$0
+
+usage() {
+  cat << EOF >&2
+Usage: $PROGNAME [--no-cleanup] [--git-add] <project path>
+
+--no-cleanup : Disable the backup/cleanup procedure
+--git-add    : Stage changed files before generating the docs
+EOF
+  exit 0
+}
+
+cleanup=true
+gitadd=false
+
+while [ "${1:-}" != "" ]; do
+  case "$1" in
+    "-h" | "--help")
+      usage
+      ;;
+    "--no-cleanup")
+      cleanup=false
+      shift
+      ;;
+    "--git-add")
+      gitadd=true
+      shift
+      ;;
+    *)
+      break
+      ;;
+  esac
+done
+
+if [ $# -eq 0 ]
+  then
+    echo "Project directory is required"
+    exit 1
+fi
 
 # install libraries for the doc-builder
 pip install -r ./requirements.txt || exit 1
@@ -26,12 +68,16 @@ fi
 rm -rf $1/docs/build
 
 function cleanup {
-  echo "Cleaning up"
-  # Restore the original docs
-  rm -rf $1/docs/
-  mv $1/docs.old/ $1/docs/
+  if [ $cleanup = true ]; then
+    echo "Cleaning up"
+    # Restore the original docs
+    rm -rf $1/docs/
+    mv $1/docs.old/ $1/docs/
+  fi
+  
   # Give read and write permissions to the docs folder, as docker root take ownership of 
   # the files
+  echo "Fixing permissions"
   chmod -R a+rw $1/docs
 }
 
@@ -41,8 +87,18 @@ function error_exit {
   exit 1
 }
 
-# Backing up the docs folder
-cp -r $1/docs/ $1/docs.old/ || exit 1
+if [ $gitadd = true ]; then
+  echo "Staging changed files"
+  cd $1 && git add docs
+  cd -
+fi
+
+if [ $cleanup = true ]; then
+  echo "Creating backup"
+  
+  # Backing up the docs folder
+  cp -r $1/docs/ $1/docs.old/ || exit 1
+fi
 
 # syncing ivy folder with the doc-builder folder
 rsync -rav docs/ $1/docs/ || error_exit $1
@@ -52,7 +108,9 @@ sphinx-build -b html $1/docs $1/docs/build || error_exit $1
 # Disable Jekyll in GitHub pages
 touch $1/docs/build/.nojekyll
 
-# Move the build to docs.old
-mv $1/docs/build $1/docs.old/build || error_exit $1
+if [ $cleanup = true ]; then
+  # Move the build to docs.old
+  mv $1/docs/build $1/docs.old/build || error_exit $1
+fi
 
 cleanup $1
